Self-Guided Learning/python/day_07/examples/simple_drf/pyproject.toml
[tool.poetry]
name = "simple-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "simple_drf" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
djangorestframework = "^3.15.2"
coreapi = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
